fix: unit32 eb / operator struct

+++ contracts/test/SSVNetworkUpgrade.sol
@@ -393,7 +393,7 @@ contract SSVNetworkUpgrade is
         address clusterOwner,
         uint64[] calldata operatorIds,
         ISSVNetworkCore.Cluster memory cluster,
-        uint256 effectiveBalance,
+        uint32 effectiveBalance,
         bytes32[] calldata merkleProof
     ) external override {
         // TODO _delegate(SSVStorage.load().ssvContracts[SSVModules.SSV_CLUSTERS]);
